+++ src/tags_int.cpp
@@ -2624,7 +2624,7 @@ uint16_t tagNumber(const std::string& tagName, IfdId ifdId) {
     return ti->tag_;
   if (!isHex(tagName, 4, "0x"))
     throw Error(ErrorCode::kerInvalidTag, tagName, ifdId);
-  return std::stoi(tagName, nullptr, 16);
+  return static_cast<uint16_t>(std::stoi(tagName, nullptr, 16));
 }  // tagNumber
 
 std::ostream& printInt64(std::ostream& os, const Value& value, const ExifData*) {
